+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   and added `fcft_font.max_advance.y`.
 * API: renamed `fcft_font.space_x_advance` to
   `fcft_font.space_advance.x` and added `fcft_font.space_advance.y`.
+* API: renamed `fcft_glyph_for_wc()` to `fcft_glyph_rasterize()`.
 
 
 ### Deprecated